Python 10 is working
Using the same algorithm as the scheme program, surprisingly the Python script will run in under five seconds. I think the cause is the scheme script having to attach each prime at the beginning of the list rather than the end. As scheme goes through each number, it is dividing larger and larger numbers by large primes to begin with rather than starting with 2, 3, 5....